third_party/blink/web_tests/ppapi/plugins/change-widget-and-click-crash-expected.txt

CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: initializing
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(true)
CONSOLE MESSAGE: Blink Test Plugin: MouseDown at (151,76)
CONSOLE MESSAGE: Blink Test Plugin: DidChangeFocus(false)
PASS
